requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
pyairtable==2.3.5
openai>=1.0,<2.0
cloudinary==1.41.0
reportlab==4.2.5
